Feat: Prepare existing models for spring security and move to model/auth package
