+++ BackEnd/src/main/java/org/sds/elevateconnect/controller/PermissionsController.java
@@ -1,6 +1,6 @@
 package org.sds.elevateconnect.controller;
 
-import org.sds.elevateconnect.model.Permission;
+import org.sds.elevateconnect.model.auth.Permission;
 import org.sds.elevateconnect.model.Result;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
