--- BackEnd/src/main/java/org/sds/elevateconnect/dto/SignupRequest.java
@@ -1,16 +1,0 @@
-package org.sds.elevateconnect.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class SignupRequest {
-    private String email;
-    private String firstName;
-    private String lastName;
-    private String password;
-    private String inviteCode;
-}
