--- BackEnd/src/main/java/org/sds/elevateconnect/model/User.java
@@ -1,26 +1,0 @@
-package org.sds.elevateconnect.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class User {
-    private Integer id;
-    private Integer communityId;
-    private String firstName;
-    private String lastName;
-    private String email;
-    private String password;
-    private UserRole role;
-    private String country;
-    private LocalDateTime createTime;
-
-    public String getFullName() {
-        return firstName + ' ' + lastName;
-    }
-}
