+++ BackEnd/src/main/java/org/sds/elevateconnect/service/InviteCodeService.java
@@ -3,8 +3,7 @@ package org.sds.elevateconnect.service;
 import org.sds.elevateconnect.exceptions.InviteCodeException;
 import org.sds.elevateconnect.mapper.InviteCodeMapper;
 import org.sds.elevateconnect.model.InviteCode;
-import org.sds.elevateconnect.model.Result;
-import org.sds.elevateconnect.model.UserRole;
+import org.sds.elevateconnect.model.auth.UserRole;
 import org.sds.elevateconnect.service.interfaces.IInviteCodeService;
 import org.sds.elevateconnect.utils.CodeGenerator;
 import org.springframework.beans.factory.annotation.Autowired;
